requirements.txt
Django==1.8.2
argparse==1.2.1
django-ckeditor==4.4.8
django-grappelli==2.6.5
wsgiref==0.1.2
